feat: create new class CurrencyExchangeController and transfer logic from class Programm into this class
