+++ CurrencyExchangeTests/UnitTest1.cs
@@ -1,3 +1,4 @@
+using CurrencyExchange;
 using CurrencyExchangeApp;
 using System;
 using Xunit;
@@ -7,18 +8,19 @@ namespace CurrencyExchangeTests
     public class UnitTest1
     {
         [Fact]
-        public void Program_WhenPathOrTextIsNull_Return_Exception()
+        public void CurrencyExchangeController_WhenPathOrTextIsNull_Return_Exception()
         {
             // Arrange
+            var controller = new CurrencyExchangeController();
             string path = "  ";
             string text = null;
-            var ag = new ArgumentNullException();
+            var arg = new ArgumentNullException();
 
             // Act
-            var ex = Assert.ThrowsAsync<ArgumentNullException>(() => Program.SaveAsync(path, text)).GetAwaiter().GetResult();
+            var ex = Assert.ThrowsAsync<ArgumentNullException>(() => CurrencyExchangeController.SaveAsync(path, text)).GetAwaiter().GetResult();
 
             // Assert
-            Assert.Equal(ex.Message, ag.Message);
+            Assert.Equal(ex.GetType(), arg.GetType());
         }
     }
 }
