+++ src/CurrencyExchange/CurrencyExchangeApp/CurrencyExchangeApp.csproj
@@ -6,10 +6,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\CurrencyExchange\CurrencyExchange.csproj" />
   </ItemGroup>
